pyproject.toml
[tool.poetry]
name = "django-rest-ecommerce"
version = "0.1.0"
description = ""
authors = ["Zihad <zihadmahiuddin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
